LayoutTests/imported/w3c/IndexedDB-private-browsing/idbcursor_delete_index5-expected.txt


PASS IDBCursor.delete() - index - throw InvalidStateError when the cursor is being iterated
